refactor: migrate presets to party-display-core
