+++ app/src-tauri/party-display-core/src/lib.rs
@@ -3,3 +3,4 @@ pub mod system;
 pub mod smtc;
 pub mod dlna;
 pub mod local_audio;
+pub mod presets;
